+++ src/stacktrace_win-inl.h
@@ -5,7 +5,7 @@
 _START_GOOGLE_NAMESPACE_
 
 // If you change this function, also change GetStackFrames below.
-int GetStackTrace(void** result, int max_depth, int skip_count) {
+size_t GetStackTrace(void** result, size_t max_depth, int skip_count) {
 // NOTE(phl): This was tuned for the static library with VS2013.
 #ifdef _DEBUG
   skip_count += 3;
@@ -19,7 +19,7 @@ int GetStackTrace(void** result, int max_depth, int skip_count) {
 
   if (size > max_depth)
     size = max_depth;
-  for (int i = 0; i < size; i++)
+  for (size_t i = 0; i < size; i++)
     result[i] = const_cast<void*>(addresses[i]);
 
   return size;
